pyproject.toml
[project]
name = "abnormal-prediction-in-chest-x-ray"
version = "0.1.0"
description = "Chest X-ray abnormality detection with image processing filters and AI-powered disease detection"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pillow>=10.1.0",
    "numpy>=1.26.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
backend = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "ultralytics>=8.0.0",
    "pydantic>=2.5.0",
]
frontend = [
    "streamlit>=1.28.0",
    "requests>=2.31.0",
]
training = [
    "roboflow>=1.1.0",
    "wandb>=0.16.0",
    "tqdm>=4.66.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.27.0",
]
dev = [
    "ruff>=0.1.0",
]
all = [
    "abnormal-prediction-in-chest-x-ray[backend,frontend,training,dev]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.hatch.build.targets.wheel]
packages = ["abnormal_prediction_in_chest_x_ray"]
